CEO-351 Reorganize SVGs.
