+++ src/js/collection.js
@@ -17,7 +17,7 @@ import {CollapsibleTitle} from "./components/FormComponents";
 import Modal from "./components/Modal";
 import Select from "./components/Select";
 import Switch from "./components/Switch";
-import {ButtonSvgIcon} from "./components/SvgIcon";
+import {ButtonSvgIcon} from "./components/svg/SvgIcon";
 
 import {UnicodeIcon, getQueryString, safeLength, isNumber, invertColor, asPercentage, isArray} from "./utils/generalUtils";
 import {getProjectPreferences, setProjectPreferences} from "./utils/preferences";
