--- src/js/components/SvgIcon.js
@@ -1,314 +1,0 @@
-import React from "react";
-
-const helpIcon = (
-    <svg viewBox="0 0 16 16">
-        <title/>
-        <defs/>
-        <g fill="none" fillRule="evenodd" id="Icons with numbers" stroke="none" strokeWidth="1">
-            <g id="Group" style={{fill: "currentColor"}} transform="translate(-48.000000, -432.000000)">
-                <path d="M54.8796844,443.0591 L54.8796844,445 L57.2307692,445 L57.2307692,443.0591 Z M56,448 C51.5817218,448 48,444.418278 48,440 C48,435.581722 51.5817218,432 56,432 C60.4182782,432 64,435.581722 64,440 C64,444.418278 60.4182782,448 56,448 Z M53.5700197,435.51041 C52.5864514,436.043208 52.0631167,436.947609 52,438.22364 L54.2800789,438.22364 C54.2800789,437.852024 54.4076253,437.493845 54.6627219,437.149093 C54.9178185,436.804341 55.3504243,436.631968 55.9605523,436.631968 C56.5811997,436.631968 57.0085458,436.771881 57.2426036,437.051713 C57.4766613,437.331544 57.5936884,437.641592 57.5936884,437.981867 C57.5936884,438.277369 57.4884955,438.548241 57.2781065,438.794493 L56.8205128,439.190732 L56.2445759,439.573539 C55.6765258,439.949633 55.3241295,440.282067 55.1873767,440.570853 C55.0506239,440.859639 54.9664696,441.382356 54.9349112,442.139019 L57.0650888,442.139019 C57.0703485,441.780835 57.1045362,441.516679 57.1676529,441.346541 C57.2675876,441.077903 57.4700839,440.842849 57.7751479,440.64137 L58.3353057,440.271995 C58.9033559,439.895901 59.28731,439.586972 59.4871795,439.345198 C59.8290615,438.946718 60,438.456461 60,437.874412 C60,436.925225 59.6068415,436.208867 58.8205128,435.725319 C58.0341841,435.241771 57.0466858,435 55.8579882,435 C54.9533157,435 54.1906671,435.170135 53.5700197,435.51041 Z M53.5700197,435.51041" id="Oval 318"/>
-            </g>
-        </g>
-    </svg>
-);
-
-const rightDoubleIcon = (
-    <svg
-        style={{verticalAlign: "initial"}}
-        viewBox="0 0 26 26"
-    >
-        <g>
-            <polygon fill="#231F20" points="2.049,0.58 -0.035,2.664 10.801,13.5 -0.035,24.336 2.049,26.42 14.969,13.5  "/>
-            <polygon fill="#231F20" points="13.049,0.58 10.965,2.664 21.801,13.5 10.965,24.336 13.049,26.42 25.969,13.5  "/>
-        </g>
-    </svg>
-);
-
-const leftDoubleIcon = (
-    <svg
-        enableBackground="new 0 0 26 26"
-        id="Layer_1"
-        style={{verticalAlign: "initial"}}
-        version="1.1"
-        viewBox="0 0 26 26"
-    >
-        <g>
-            <polygon fill="#231F20" points="23.885,0.58 25.969,2.664 15.133,13.5 25.969,24.336 23.885,26.42 10.965,13.5  "/>
-            <polygon fill="#231F20" points="12.885,0.58 14.969,2.664 4.133,13.5 14.969,24.336 12.885,26.42 -0.035,13.5  "/>
-        </g>
-    </svg>
-);
-
-const closeIcon = (
-    <svg viewBox="0 0 512 512">
-        <path
-            d="M256,0C114.508,0,0,114.497,0,256c0,141.493,114.497,256,256,256c141.492,0,256-114.497,256-256
-            C512,114.507,397.503,0,256,0z M256,472c-119.384,0-216-96.607-216-216c0-119.385,96.607-216,216-216
-            c119.384,0,216,96.607,216,216C472,375.385,375.393,472,256,472z"
-        />
-        <path
-            d="M343.586,315.302L284.284,256l59.302-59.302c7.81-7.81,7.811-20.473,0.001-28.284c-7.812-7.811-20.475-7.81-28.284,0
-            L256,227.716l-59.303-59.302c-7.809-7.811-20.474-7.811-28.284,0c-7.81,7.811-7.81,20.474,0.001,28.284L227.716,256
-            l-59.302,59.302c-7.811,7.811-7.812,20.474-0.001,28.284c7.813,7.812,20.476,7.809,28.284,0L256,284.284l59.303,59.302
-            c7.808,7.81,20.473,7.811,28.284,0C351.398,335.775,351.397,323.112,343.586,315.302z"
-        />
-    </svg>
-);
-
-const drawIcon = (
-    <svg viewBox="0 0 283.093 283.093">
-        <g>
-            <path
-                d="M271.315,54.522L218.989,2.196c-2.93-2.928-7.678-2.928-10.607,0L78.274,132.303c-1.049,1.05-1.764,2.388-2.053,3.843
-                l-12.964,65.29c-0.487,2.456,0.282,4.994,2.053,6.765c1.421,1.42,3.334,2.196,5.304,2.196c0.485,0,0.975-0.047,1.461-0.144
-                l65.29-12.964c1.456-0.289,2.793-1.004,3.843-2.053L271.315,65.129C274.244,62.2,274.244,57.452,271.315,54.522z M83.182,178.114
-                l6.776-34.127l39.566,39.566l-34.127,6.776L83.182,178.114z"
-            />
-            <path
-                d="M205.912,227.066c-71.729-30.029-118.425,19.633-132.371,27.175c-17.827,9.641-42.941,20.97-48.779,1.358
-                c-3.522-11.832,15.521-24.479,28.131-28.42c9.2-2.876,5.271-17.358-3.988-14.465c-19.582,6.121-42.948,22.616-38.851,45.839
-                c3.044,17.256,24.67,32.995,66.368,11.114c30.308-15.902,50.897-48.84,114.733-31.783c20.969,5.602,37.92,19.27,45.178,40.057
-                c3.168,9.07,17.662,5.169,14.465-3.988C243.062,251.799,227.411,236.067,205.912,227.066z"
-            />
-        </g>
-    </svg>
-);
-
-const pointIcon = (
-    <svg viewBox="0 0 341.333 341.333">
-        <path
-            d="m 170.667,93.86685 c -42.41565,0 -76.80015,34.3845 -76.80015,76.80015 0,42.41565 34.3845,76.80015 76.80015,76.80015
-            42.41565,0 76.80015,-34.3845 76.80015,-76.80015 0,-42.41565 -34.38495,-76.80015 -76.80015,-76.80015 z m 0,134.40015 c
-            -31.8114,0 -57.6,-25.7886 -57.6,-57.6 0,-31.8114 25.7886,-57.6 57.6,-57.6 31.8114,0 57.6,25.7886 57.6,57.6 0,31.8114
-            -25.7886,57.6 -57.6,57.6 z"
-        />
-        <path
-            d="m 170.667,109.22688 c -33.93252,0 -61.44012,27.5076 -61.44012,61.44012 0,33.93252 27.5076,61.44012 61.44012,61.44012
-            33.93252,0 61.44012,-27.5076 61.44012,-61.44012 0,-33.93252 -27.50796,-61.44012 -61.44012,-61.44012 z m 0,107.52012 c
-            -25.44912,0 -46.08,-20.63088 -46.08,-46.08 0,-25.44912 20.63088,-46.08 46.08,-46.08 25.44912,0 46.08,20.63088 46.08,46.08
-            0,25.44912 -20.63088,46.08 -46.08,46.08 z"
-        />
-    </svg>
-);
-
-const lineStringIcon = (
-    <svg viewBox="0 0 512 512">
-        <path
-            d="m512 98.857v-98.857h-98.857v77.645l-128.927 128.926h-77.644v77.644l-128.928
-            128.928h-77.644v98.857h98.857v-77.645l128.927-128.927h77.645v-77.645l128.926-128.926zm-68.857-68.857h38.857v38.857h-38.857zm-374.286
-            452h-38.857v-38.857h38.857zm206.572-206.571h-38.857v-38.857h38.857z"
-        />
-    </svg>
-);
-
-const polygonIcon = (
-    <svg viewBox="0 0 512 512">
-        <path
-            d="m497.027344 209.589844h-9.8125l-69.691406-119.765625h9.640624c8.269532 0 14.972657-6.703125
-            14.972657-14.96875v-59.882813c0-8.269531-6.703125-14.972656-14.972657-14.972656h-59.882812c-8.265625
-            0-14.96875 6.703125-14.96875
-            14.972656v16.253906h-192.625v-16.253906c0-8.269531-6.703125-14.972656-14.96875-14.972656h-59.882812c-8.269532
-            0-14.972657 6.703125-14.972657 14.972656v59.882813c0 8.265625 6.703125 14.96875 14.972657 14.96875h7.925781l-69.148438
-            119.765625h-8.640625c-8.269531 0-14.972656 6.703125-14.972656 14.972656v59.882812c0 8.265626 6.703125 14.96875
-            14.972656 14.96875h8.640625l70.300781 121.761719h-9.078124c-8.269532 0-14.972657 6.703125-14.972657 14.972657v59.882812c0
-            8.269531 6.703125 14.96875 14.972657 14.96875h59.882812c8.265625 0 14.96875-6.699219
-            14.96875-14.96875v-18.253906h192.625v18.253906c0 8.269531 6.703125 14.96875 14.96875 14.96875h59.882812c8.269532
-            0 14.972657-6.699219
-            14.972657-14.96875v-59.882812c0-8.269532-6.703125-14.972657-14.972657-14.972657h-10.074218l70.296875-121.761719h9.640625c8.269531
-            0 14.972656-6.703124 14.972656-14.96875v-59.882812c0-8.269531-6.703125-14.972656-14.972656-14.972656zm-14.96875
-            29.941406v29.941406h-29.941406v-29.941406zm-99.804688-209.589844h29.941406v29.941406h-29.941406zm-282.449218
-            0h29.941406v29.941406h-29.941406zm-69.863282 239.53125v-29.941406h29.941406v29.941406zm99.804688
-            211.585938h-29.941406v-29.941406h29.941406zm282.449218 0h-29.941406v-29.941406h29.941406zm-29.679687-59.882813h-15.234375c-8.265625
-            0-14.96875 6.703125-14.96875
-            14.972657v11.6875h-192.625v-11.6875c0-8.269532-6.703125-14.972657-14.96875-14.972657h-16.230469l-70.300781-121.761719h16.667969c8.265625
-            0 14.96875-6.703124
-            14.96875-14.96875v-59.882812c0-8.269531-6.703125-14.972656-14.96875-14.972656h-16.667969l69.148438-119.765625h17.382812c8.265625
-            0 14.96875-6.703125 14.96875-14.96875v-13.6875h192.625v13.6875c0 8.265625 6.703125 14.96875 14.96875 14.96875h15.597656l69.695313
-            119.765625h-15.429688c-8.265625 0-14.96875 6.703125-14.96875 14.972656v59.882812c0 8.265626 6.703125 14.96875 14.96875
-            14.96875h15.671875zm0 0"
-        />
-    </svg>
-);
-
-const questionIcon = (
-    <svg viewBox="0 0 31.357 31.357">
-        <path
-            d="M15.255,0c5.424,0,10.764,2.498,10.764,8.473c0,5.51-6.314,7.629-7.67,9.62c-1.018,1.481-0.678,3.562-3.475,3.562
-            c-1.822,0-2.712-1.482-2.712-2.838c0-5.046,7.414-6.188,7.414-10.343c0-2.287-1.522-3.643-4.066-3.643
-            c-5.424,0-3.306,5.592-7.414,5.592c-1.483,0-2.756-0.89-2.756-2.584C5.339,3.683,10.084,0,15.255,0z M15.044,24.406
-            c1.904,0,3.475,1.566,3.475,3.476c0,1.91-1.568,3.476-3.475,3.476c-1.907,0-3.476-1.564-3.476-3.476
-            C11.568,25.973,13.137,24.406,15.044,24.406z"
-        />
-    </svg>
-);
-
-const checkIcon = (
-    <svg
-        style={{verticalAlign: "initial"}}
-        viewBox="0 0 352.62 352.62"
-    >
-        <path
-            d="M337.222,22.952c-15.912-8.568-33.66,7.956-44.064,17.748c-23.867,23.256-44.063,50.184-66.708,74.664
-            c-25.092,26.928-48.348,53.856-74.052,80.173c-14.688,14.688-30.6,30.6-40.392,48.96c-22.032-21.421-41.004-44.677-65.484-63.648
-            c-17.748-13.464-47.124-23.256-46.512,9.18c1.224,42.229,38.556,87.517,66.096,116.28c11.628,12.24,26.928,25.092,44.676,25.704
-            c21.42,1.224,43.452-24.48,56.304-38.556c22.645-24.48,41.005-52.021,61.812-77.112c26.928-33.048,54.468-65.485,80.784-99.145
-            C326.206,96.392,378.226,44.983,337.222,22.952z M26.937,187.581c-0.612,0-1.224,0-2.448,0.611
-            c-2.448-0.611-4.284-1.224-6.732-2.448l0,0C19.593,184.52,22.653,185.132,26.937,187.581z"
-        />
-    </svg>
-);
-
-const ruleIcon = (
-    <svg viewBox="0 0 438.891 438.891">
-        <path
-            d="M347.968,57.503h-39.706V39.74c0-5.747-6.269-8.359-12.016-8.359h-30.824c-7.314-20.898-25.6-31.347-46.498-31.347
-            c-20.668-0.777-39.467,11.896-46.498,31.347h-30.302c-5.747,0-11.494,2.612-11.494,8.359v17.763H90.923
-            c-23.53,0.251-42.78,18.813-43.886,42.318v299.363c0,22.988,20.898,39.706,43.886,39.706h257.045
-            c22.988,0,43.886-16.718,43.886-39.706V99.822C390.748,76.316,371.498,57.754,347.968,57.503z M151.527,52.279h28.735
-            c5.016-0.612,9.045-4.428,9.927-9.404c3.094-13.474,14.915-23.146,28.735-23.51c13.692,0.415,25.335,10.117,28.212,23.51
-            c0.937,5.148,5.232,9.013,10.449,9.404h29.78v41.796H151.527V52.279z M370.956,399.185c0,11.494-11.494,18.808-22.988,18.808
-            H90.923c-11.494,0-22.988-7.314-22.988-18.808V99.822c1.066-11.964,10.978-21.201,22.988-21.42h39.706v26.645
-            c0.552,5.854,5.622,10.233,11.494,9.927h154.122c5.98,0.327,11.209-3.992,12.016-9.927V78.401h39.706
-            c12.009,0.22,21.922,9.456,22.988,21.42V399.185z"
-        />
-        <path
-            d="M179.217,233.569c-3.919-4.131-10.425-4.364-14.629-0.522l-33.437,31.869l-14.106-14.629
-            c-3.919-4.131-10.425-4.363-14.629-0.522c-4.047,4.24-4.047,10.911,0,15.151l21.42,21.943c1.854,2.076,4.532,3.224,7.314,3.135
-            c2.756-0.039,5.385-1.166,7.314-3.135l40.751-38.661c4.04-3.706,4.31-9.986,0.603-14.025
-            C179.628,233.962,179.427,233.761,179.217,233.569z"
-        />
-        <path
-            d="M329.16,256.034H208.997c-5.771,0-10.449,4.678-10.449,10.449s4.678,10.449,10.449,10.449H329.16
-            c5.771,0,10.449-4.678,10.449-10.449S334.931,256.034,329.16,256.034z"
-        />
-        <path
-            d="M179.217,149.977c-3.919-4.131-10.425-4.364-14.629-0.522l-33.437,31.869l-14.106-14.629
-            c-3.919-4.131-10.425-4.364-14.629-0.522c-4.047,4.24-4.047,10.911,0,15.151l21.42,21.943c1.854,2.076,4.532,3.224,7.314,3.135
-            c2.756-0.039,5.385-1.166,7.314-3.135l40.751-38.661c4.04-3.706,4.31-9.986,0.603-14.025
-            C179.628,150.37,179.427,150.169,179.217,149.977z"
-        />
-        <path
-            d="M329.16,172.442H208.997c-5.771,0-10.449,4.678-10.449,10.449s4.678,10.449,10.449,10.449H329.16
-            c5.771,0,10.449-4.678,10.449-10.449S334.931,172.442,329.16,172.442z"
-        />
-        <path
-            d="M179.217,317.16c-3.919-4.131-10.425-4.363-14.629-0.522l-33.437,31.869l-14.106-14.629
-            c-3.919-4.131-10.425-4.363-14.629-0.522c-4.047,4.24-4.047,10.911,0,15.151l21.42,21.943c1.854,2.076,4.532,3.224,7.314,3.135
-            c2.756-0.039,5.385-1.166,7.314-3.135l40.751-38.661c4.04-3.706,4.31-9.986,0.603-14.025
-            C179.628,317.554,179.427,317.353,179.217,317.16z"
-        />
-        <path
-            d="M329.16,339.626H208.997c-5.771,0-10.449,4.678-10.449,10.449s4.678,10.449,10.449,10.449H329.16
-            c5.771,0,10.449-4.678,10.449-10.449S334.931,339.626,329.16,339.626z"
-        />
-    </svg>
-);
-
-const leftArrowIcon = (
-    <svg transform="translate(-2 0)" viewBox="0 0 512.2 512.2" x="0px" y="0px">
-        <path
-            d="M501.534,160.111H234.867V42.778c0.011-5.891-4.757-10.675-10.648-10.686 c-2.84-0.005-5.565,1.123-7.571,3.134L3.123,248.559
-            c-4.164,4.165-4.164,10.917,0,15.083l213.525,213.333 c4.16,4.171,10.914,4.179,15.085,0.019c2.006-2.001,3.133-4.717,3.134-7.55V352.111h266.667
-            c5.891,0,10.667-4.776,10.667-10.667 V170.778C512.2,164.886,507.425,160.111,501.534,160.111z"
-        />
-    </svg>
-);
-
-const rightArrowIcon = (
-    <svg viewBox="0 0 512.19 512.19" x="0px" y="0px">
-        <path
-            d="M295.445,35.381c-4.218-4.112-10.972-4.026-15.084,0.192c-1.921,1.971-3.006,4.608-3.028,7.36 v117.333H10.667
-            C4.776,160.266,0,165.042,0,170.933V341.6c0,5.891,4.776,10.667,10.667,10.667h266.667v116.885 c-0.011,5.891,4.757,10.675,10.648,10.686
-            c2.84,0.005,5.565-1.123,7.571-3.134L509.056,263.37c4.171-4.16,4.179-10.914,0.019-15.085
-            c-0.006-0.006-0.013-0.013-0.019-0.019L295.445,35.381z"
-        />
-    </svg>
-);
-
-const editIcon = (
-    <svg viewBox="0 0 512 512" x="0px" y="0px">
-        <rect
-            height="318.717"
-            transform="matrix(-0.7073 -0.7069 0.7069 -0.7073 233.6377 627.3448)"
-            width="160.05"
-            x="166.669"
-            y="105.945"
-        />
-        <path
-            d="M502.625,77.251L434.749,9.375C428.539,3.164,420.413,0.041,412.283,0h-0.338c-8.131,0.041-16.246,3.164-22.441,9.375
-            l-45.312,45.251l113.183,113.121l45.251-45.251c6.211-6.211,9.334-14.326,9.375-22.456v-0.338
-            C511.959,91.576,508.836,83.461,502.625,77.251z"
-        />
-        <polygon points="0,512 147.072,478.049 33.889,364.928"/>
-    </svg>
-);
-
-const plus = (
-    <svg viewBox="0 0 512 512">
-        <path d="m467 211h-166v-166c0-24.853-20.147-45-45-45s-45 20.147-45 45v166h-166c-24.853 0-45 20.147-45 45s20.147 45 45 45h166v166c0 24.853 20.147 45 45 45s45-20.147 45-45v-166h166c24.853 0 45-20.147 45-45s-20.147-45-45-45z"/>
-    </svg>
-);
-
-const minus = (
-    <svg viewBox="0 0 512 512" x="0px" y="0px">
-        <path
-            d="M467,211L45,211C20.147,211 0,231.147 0,256C0,280.853 20.147,301 45,301L467,301C491.853,301 512,280.853 512,256C512,231.147 491.853,211 467,211Z"
-            style={{"fill-rule": "nonzero"}}
-        />
-    </svg>
-);
-
-const iconMap = {
-    "check": checkIcon,
-    "close": closeIcon,
-    "draw": drawIcon,
-    "edit": editIcon,
-    "help": helpIcon,
-    "leftArrow": leftArrowIcon,
-    "leftDouble": leftDoubleIcon,
-    "lineString": lineStringIcon,
-    "minus": minus,
-    "plus": plus,
-    "point": pointIcon,
-    "polygon": polygonIcon,
-    "question": questionIcon,
-    "rightArrow": rightArrowIcon,
-    "rightDouble": rightDoubleIcon,
-    "rule": ruleIcon
-};
-
-export default function SvgIcon({icon, color, size}) {
-    return (
-        <div
-            style={{
-                color: color || "black",
-                fill: color || "black",
-                height: size,
-                maxHeight: size,
-                maxWidth: size,
-                cursor: "pointer",
-                width: size
-            }}
-        >
-            {iconMap[icon]}
-        </div>
-    );
-}
-
-export function ButtonSvgIcon({icon, size}) {
-    return (
-        <div
-            className="svg-icon"
-            style={{
-                fill: "currentColor",
-                height: size,
-                maxHeight: size,
-                maxWidth: size,
-                cursor: "pointer",
-                width: size
-            }}
-        >
-            {iconMap[icon]}
-        </div>
-    );
-}
