+++ src/js/components/svg/editIcon.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+export const editIcon = (
+    <svg viewBox="0 0 512 512" x="0px" y="0px">
+        <rect
+            height="318.717"
+            transform="matrix(-0.7073 -0.7069 0.7069 -0.7073 233.6377 627.3448)"
+            width="160.05"
+            x="166.669"
+            y="105.945"
+        />
+        <path
+            d="M502.625,77.251L434.749,9.375C428.539,3.164,420.413,0.041,412.283,0h-0.338c-8.131,0.041-16.246,3.164-22.441,9.375l-45.312,45.251l113.183,113.121l45.251-45.251c6.211-6.211,9.334-14.326,9.375-22.456v-0.338C511.959,91.576,508.836,83.461,502.625,77.251z"
+        />
+        <polygon points="0,512 147.072,478.049 33.889,364.928"/>
+    </svg>
+);
