+++ src/js/components/svg/leftArrowIcon.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const leftArrowIcon = (
+    <svg transform="translate(-2 0)" viewBox="0 0 512.2 512.2" x="0px" y="0px">
+        <path
+            d="M501.534,160.111H234.867V42.778c0.011-5.891-4.757-10.675-10.648-10.686 c-2.84-0.005-5.565,1.123-7.571,3.134L3.123,248.559c-4.164,4.165-4.164,10.917,0,15.083l213.525,213.333 c4.16,4.171,10.914,4.179,15.085,0.019c2.006-2.001,3.133-4.717,3.134-7.55V352.111h266.667c5.891,0,10.667-4.776,10.667-10.667 V170.778C512.2,164.886,507.425,160.111,501.534,160.111z"
+        />
+    </svg>
+);
