+++ src/js/components/svg/lineStringIcon.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const lineStringIcon = (
+    <svg viewBox="0 0 512 512">
+        <path
+            d="m512 98.857v-98.857h-98.857v77.645l-128.927 128.926h-77.644v77.644l-128.928128.928h-77.644v98.857h98.857v-77.645l128.927-128.927h77.645v-77.645l128.926-128.926zm-68.857-68.857h38.857v38.857h-38.857zm-374.286452h-38.857v-38.857h38.857zm206.572-206.571h-38.857v-38.857h38.857z"
+        />
+    </svg>
+);
