+++ src/js/components/svg/plus.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const plus = (
+    <svg viewBox="0 0 512 512">
+        <path d="m467 211h-166v-166c0-24.853-20.147-45-45-45s-45 20.147-45 45v166h-166c-24.853 0-45 20.147-45 45s20.147 45 45 45h166v166c0 24.853 20.147 45 45 45s45-20.147 45-45v-166h166c24.853 0 45-20.147 45-45s-20.147-45-45-45z"/>
+    </svg>
+);
