+++ src/js/components/svg/pointIcon.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const pointIcon = (
+    <svg viewBox="0 0 341.333 341.333">
+        <path
+            d="m 170.667,93.86685 c -42.41565,0 -76.80015,34.3845 -76.80015,76.80015 0,42.41565 34.3845,76.80015 76.80015,76.8001542.41565,0 76.80015,-34.3845 76.80015,-76.80015 0,-42.41565 -34.38495,-76.80015 -76.80015,-76.80015 z m 0,134.40015 c-31.8114,0 -57.6,-25.7886 -57.6,-57.6 0,-31.8114 25.7886,-57.6 57.6,-57.6 31.8114,0 57.6,25.7886 57.6,57.6 0,31.8114-25.7886,57.6 -57.6,57.6 z"
+        />
+        <path
+            d="m 170.667,109.22688 c -33.93252,0 -61.44012,27.5076 -61.44012,61.44012 0,33.93252 27.5076,61.44012 61.44012,61.4401233.93252,0 61.44012,-27.5076 61.44012,-61.44012 0,-33.93252 -27.50796,-61.44012 -61.44012,-61.44012 z m 0,107.52012 c-25.44912,0 -46.08,-20.63088 -46.08,-46.08 0,-25.44912 20.63088,-46.08 46.08,-46.08 25.44912,0 46.08,20.63088 46.08,46.080,25.44912 -20.63088,46.08 -46.08,46.08 z"
+        />
+    </svg>
+);
