+++ src/js/components/svg/polygonIcon.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const polygonIcon = (
+    <svg viewBox="0 0 512 512">
+        <path
+            d="m497.027344 209.589844h-9.8125l-69.691406-119.765625h9.640624c8.269532 0 14.972657-6.70312514.972657-14.96875v-59.882813c0-8.269531-6.703125-14.972656-14.972657-14.972656h-59.882812c-8.2656250-14.96875 6.703125-14.9687514.972656v16.253906h-192.625v-16.253906c0-8.269531-6.703125-14.972656-14.96875-14.972656h-59.882812c-8.2695320-14.972657 6.703125-14.972657 14.972656v59.882813c0 8.265625 6.703125 14.96875 14.972657 14.96875h7.925781l-69.148438119.765625h-8.640625c-8.269531 0-14.972656 6.703125-14.972656 14.972656v59.882812c0 8.265626 6.703125 14.9687514.972656 14.96875h8.640625l70.300781 121.761719h-9.078124c-8.269532 0-14.972657 6.703125-14.972657 14.972657v59.882812c08.269531 6.703125 14.96875 14.972657 14.96875h59.882812c8.265625 0 14.96875-6.69921914.96875-14.96875v-18.253906h192.625v18.253906c0 8.269531 6.703125 14.96875 14.96875 14.96875h59.882812c8.2695320 14.972657-6.69921914.972657-14.96875v-59.882812c0-8.269532-6.703125-14.972657-14.972657-14.972657h-10.074218l70.296875-121.761719h9.640625c8.2695310 14.972656-6.703124 14.972656-14.96875v-59.882812c0-8.269531-6.703125-14.972656-14.972656-14.972656zm-14.9687529.941406v29.941406h-29.941406v-29.941406zm-99.804688-209.589844h29.941406v29.941406h-29.941406zm-282.4492180h29.941406v29.941406h-29.941406zm-69.863282 239.53125v-29.941406h29.941406v29.941406zm99.804688211.585938h-29.941406v-29.941406h29.941406zm282.449218 0h-29.941406v-29.941406h29.941406zm-29.679687-59.882813h-15.234375c-8.2656250-14.96875 6.703125-14.9687514.972657v11.6875h-192.625v-11.6875c0-8.269532-6.703125-14.972657-14.96875-14.972657h-16.230469l-70.300781-121.761719h16.667969c8.2656250 14.96875-6.70312414.96875-14.96875v-59.882812c0-8.269531-6.703125-14.972656-14.96875-14.972656h-16.667969l69.148438-119.765625h17.382812c8.2656250 14.96875-6.703125 14.96875-14.96875v-13.6875h192.625v13.6875c0 8.265625 6.703125 14.96875 14.96875 14.96875h15.597656l69.695313119.765625h-15.429688c-8.265625 0-14.96875 6.703125-14.96875 14.972656v59.882812c0 8.265626 6.703125 14.96875 14.9687514.96875h15.671875zm0 0"
+        />
+    </svg>
+);
