+++ src/js/components/svg/rightArrowIcon.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const rightArrowIcon = (
+    <svg viewBox="0 0 512.19 512.19" x="0px" y="0px">
+        <path
+            d="M295.445,35.381c-4.218-4.112-10.972-4.026-15.084,0.192c-1.921,1.971-3.006,4.608-3.028,7.36 v117.333H10.667C4.776,160.266,0,165.042,0,170.933V341.6c0,5.891,4.776,10.667,10.667,10.667h266.667v116.885 c-0.011,5.891,4.757,10.675,10.648,10.686c2.84,0.005,5.565-1.123,7.571-3.134L509.056,263.37c4.171-4.16,4.179-10.914,0.019-15.085c-0.006-0.006-0.013-0.013-0.019-0.019L295.445,35.381z"
+        />
+    </svg>
+);
