+++ src/js/components/svg/rightDoubleIcon.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+export const rightDoubleIcon = (
+    <svg
+        style={{verticalAlign: "initial"}}
+        viewBox="0 0 26 26"
+    >
+        <g>
+            <polygon fill="#231F20" points="2.049,0.58 -0.035,2.664 10.801,13.5 -0.035,24.336 2.049,26.42 14.969,13.5  "/>
+            <polygon fill="#231F20" points="13.049,0.58 10.965,2.664 21.801,13.5 10.965,24.336 13.049,26.42 25.969,13.5  "/>
+        </g>
+    </svg>
+);
